* Add unwind segue actions  * Guard against bad coding for invalid server data (at the moment caused by an incomplete EditServer implementation)  * Create and implement protocol to handle segues.  * use NSUInteger instead of uint32_t  * Add initial imll of serverdetailviewcontroller  * Rename ServerViewController to ServerActivityViewController and make ServerViewController the protocol  *

+++ Classes/ServerData.h
@@ -11,10 +11,12 @@
 @interface ServerData : NSObject
 
 + (uint32_t) serverCount;
-+ (void)     addServer: (NSArray*) serverData;
-+ (void)     deleteServerAtIndex: (uint32_t)index;
-+ (void)     setServerData: (NSArray*)serverData 
-                   atIndex: (uint32_t)index;
-+ (NSArray*) getServerDataAtIndex: (uint32_t)index;
++ (void)     addServer:             (NSArray*)      serverData;
++ (void)     deleteServerAtIndex:   (NSUInteger)    index;
+
++ (void)     setServerData:         (NSArray*)      serverData
+                   atIndex:         (NSUInteger)    index;
+
++ (NSArray*) getServerDataAtIndex:  (NSUInteger)    index;
 
 @end
